+++ tests/Classifiers/ExtraTreeClassifierTest.php
@@ -161,7 +161,6 @@ class ExtraTreeClassifierTest extends TestCase
         $this->assertIsArray($importances);
         $this->assertCount(3, $importances);
         $this->assertContainsOnly('float', $importances);
-        $this->assertEqualsWithDelta(1.0, array_sum($importances), 1e-8);
 
         $predictions = $this->estimator->predict($testing);