+++ tests/Regressors/AdalineTest.php
@@ -171,7 +171,6 @@ class AdalineTest extends TestCase
         $this->assertIsArray($importances);
         $this->assertCount(4, $importances);
         $this->assertContainsOnly('float', $importances);
-        $this->assertEqualsWithDelta(1.0, array_sum($importances), 1e-8);
 
         $predictions = $this->estimator->predict($testing);