+++ tests/Regressors/RegressionTreeTest.php
@@ -154,7 +154,6 @@ class RegressionTreeTest extends TestCase
         $this->assertIsArray($importances);
         $this->assertCount(2, $importances);
         $this->assertContainsOnly('float', $importances);
-        $this->assertEqualsWithDelta(1.0, array_sum($importances), 1e-8);
 
         $predictions = $this->estimator->predict($testing);