+++ tests/Regressors/RidgeTest.php
@@ -146,7 +146,6 @@ class RidgeTest extends TestCase
         $this->assertIsArray($importances);
         $this->assertCount(4, $importances);
         $this->assertContainsOnly('float', $importances);
-        $this->assertEqualsWithDelta(1.0, array_sum($importances), 1e-8);
 
         $predictions = $this->estimator->predict($testing);