+++ CHANGELOG.md
@@ -37,6 +37,7 @@
     - Changed Error Analysis metrics
     - Remove threat score from Multiclass Breakdown
     - Rename Labels Are Missing exception
+    - Feature importances are no longer normalized
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features